@@ -27,6 +27,9 @@ using Ae.BaoYang.Service.Client.Clients;
 using Ae.BaoYang.Service.Extension;
 using ApolloErp.Component.Http;
 using Ae.BaoYang.Service.Filters;
+using Ae.BaoYang.Service.Common.Format;
+using Microsoft.Extensions.Hosting;
+using Microsoft.AspNetCore.Http;
 
 namespace Ae.BaoYang.Service
 {
@@ -42,12 +45,19 @@ namespace Ae.BaoYang.Service
 
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddMvc()
-                .AddJsonOptions(options => { options.SerializerSettings.DateFormatString = "yyyy-MM-dd HH:mm:ss"; })
-                .SetCompatibilityVersion(CompatibilityVersion.Version_2_1).AddFluentValidation(o =>
+            //services.AddMvc()
+            //    .AddJsonOptions(options => { options.SerializerSettings.DateFormatString = "yyyy-MM-dd HH:mm:ss"; })
+            //    .SetCompatibilityVersion(CompatibilityVersion.Version_2_1).AddFluentValidation(o =>
+            //    {
+            //        o.RegisterValidatorsFromAssemblyContaining<BaoYangPartAdaptationsRequestValidator>();
+            //    });
+            services.AddControllers()
+                .AddJsonOptions(options => { options.JsonSerializerOptions.Converters.Add(new DatetimeJsonConverter("yyyy-MM-dd HH:mm:ss")); })
+                .AddFluentValidation(o =>
                 {
                     o.RegisterValidatorsFromAssemblyContaining<BaoYangPartAdaptationsRequestValidator>();
                 });
+
 
             // override modelstate
             services.Configure<ApiBehaviorOptions>(options =>
@@ -98,9 +108,10 @@ namespace Ae.BaoYang.Service
             });
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
-            if (!env.IsProduction())
+            app.UseRouting();
+            if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
 
@@ -128,11 +139,22 @@ namespace Ae.BaoYang.Service
             app.UseApolloErpCorrelationId();
 
             app.UseHttpsRedirection();
-            app.UseMvc(routes =>
+            //app.UseMvc(routes =>
+            //{
+            //    routes.MapRoute(
+            //        "default",
+            //        "{controller=Home}/{action=Index}");
+            //});
+            app.UseEndpoints(endpoints =>
             {
-                routes.MapRoute(
-                    "default",
-                    "{controller=Home}/{action=Index}");
+                // 设置默认路由
+                endpoints.MapGet("/", async context =>
+                {
+                    await context.Response.WriteAsync("Hello, World!");
+                });
+
+                // 配置控制器路由
+                endpoints.MapControllers();
             });
         }
     }
